+++ cypress/integration/executions_spec.js
@@ -31,9 +31,9 @@ describe('Dashboard Executions Page', () => {
 
       // shows a summary count of completed and failed executions
       cy.get('.overview-num__wrapper ul li')
-        .first().contains('li', 'Completed')
-        .next().contains('li', 'Failed')
-        .next().contains('li', 'Running');
+        .first().contains('li', 'Completed').contains('li', 4)
+        .next().contains('li', 'Failed').contains('li', 1)
+        .next().contains('li', 'Running').contains('li', 1);
     });
 
     it('should display the correct executions with Ids and status ', () => {