+++ cypress/integration/granules_spec.js
@@ -27,12 +27,9 @@ describe('Dashboard Granules Page', () => {
 
       // shows a summary count of completed and failed granules
       cy.get('.overview-num__wrapper ul li')
-        .first()
-        .contains('li', 'Completed')
-        .next()
-        .contains('li', 'Failed')
-        .next()
-        .contains('li', 'Running');
+        .first().contains('li', 'Completed').contains('li', 6)
+        .next().contains('li', 'Failed').contains('li', 2)
+        .next().contains('li', 'Running').contains('li', 2);
 
       // shows a list of granules
       cy.getFakeApiFixture('granules').as('granulesListFixture');