+++ cypress/integration/main_page_spec.js
@@ -71,12 +71,12 @@ describe('Dashboard Home Page', () => {
     });
 
     it('Updates start and end time components when dropdown is selected', () => {
-      const now = Date.UTC(2009, 0, 5, 13, 35, 3);
+      const now = Date.UTC(2009, 0, 5, 13, 35, 3); // 2009-01-05T13:35:03.000Z
       cy.clock(now);
       cy.get('main[class=main] section').eq(1).within(() => {
         cy.get('h3').should('have.text', 'Date and Time Range');
         cy.get('[data-cy=datetime-dropdown]').as('dateRange');
-        cy.get('@dateRange').select('Last week');
+        cy.get('@dateRange').select('1 week');
 
         cy.get('[data-cy=endDateTime]').within(() => {
           cy.get('.react-datetime-picker__inputGroup__year').should('have.value', '2009');
@@ -110,12 +110,12 @@ describe('Dashboard Home Page', () => {
     });
 
     it('should retain query parameters when moving between pages.', () => {
-      const now = Date.UTC(2015, 2, 17, 16, 0, 0);
+      const now = Date.UTC(2015, 2, 17, 16, 0, 0); // 2015-03-17T16:00:00.000Z
       cy.clock(now);
       cy.get('main[class=main] section').eq(1).within(() => {
         cy.get('h3').should('have.text', 'Date and Time Range');
         cy.get('[data-cy=datetime-dropdown]').as('dateRange');
-        cy.get('@dateRange').select('Last hour');
+        cy.get('@dateRange').select('1 hour');
 
         cy.url().should('include', 'startDateTime=201503171500');
         cy.url().should('include', 'endDateTime=201503171600');
@@ -166,6 +166,50 @@ describe('Dashboard Home Page', () => {
         cy.get('input[name=year]').should('have.value', '');
         cy.get('input[name=hour12]').should('have.value', '');
       });
+    });
+
+    it('modifies the UPDATES section and Granules Errors list as datepicker changes.', () => {
+      cy.server();
+      cy.route('GET', '/stats?*timestamp__from=1233360000000*').as('stats');
+
+      cy.get('#Errors').contains('2');
+      cy.get('#Collections').contains('1');
+      cy.get('#Granules').contains('10');
+      cy.get('#Executions').contains('6');
+
+      // Test there are values in Granule Error list
+      cy.get('[data-value="0"]').contains('FileNotFound');
+      cy.get('[data-value="1"]').contains('FileNotFound');
+
+      // This selector fails cy.get('#Ingest Rules').contains('1');
+      cy.get('.overview-num__wrapper-home > ul > :nth-child(5)').contains('1');
+
+      cy.get('[data-cy=startDateTime]').within(() => {
+        cy.get('input[name=month]').click().type(1);
+        cy.get('input[name=day]').click().type(31);
+        cy.get('input[name=year]').click().type(2009);
+        cy.get('input[name=hour12]').click().type(0);
+        cy.get('input[name=minute]').click().type(0);
+        cy.get('select[name=amPm]').select('AM');
+      });
+      cy.get('[data-cy=endDateTime]').within(() => {
+        cy.get('input[name=month]').click().type(5);
+        cy.get('input[name=day]').click().type(1);
+        cy.get('input[name=year]').click().type(2010);
+        cy.get('input[name=hour12]').click().type(0);
+        cy.get('input[name=minute]').click().type(0);
+        cy.get('select[name=amPm]').select('AM');
+      });
+
+      cy.wait('@stats');
+      cy.get('#Errors').contains('0');
+      cy.get('#Collections').contains('0');
+      cy.get('#Granules').contains('0');
+      cy.get('#Executions').contains('0');
+      cy.get('.overview-num__wrapper-home > ul > :nth-child(5)').contains('0');
+
+      // Test the Granule Error list is empty
+      cy.get('.table--wrapper > form > div > div.tbody').should('be.empty');
     });
 
     it('Logging out successfully redirects to the login screen', () => {